requirements.txt
discord.py
google-genai
google-generativeai
websockets
orjson
cryptography
aiohttp
httpx
numpy
dotenv
fernet
Pillow